Rename "Eigenlayer" to "EigenLayer"
